requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1
torchvision==0.20.1
streamlit>=1.30.0
streamlit-drawable-canvas==0.9.3
numpy<2.0.0
matplotlib
seaborn
scikit-learn
opencv-python-headless
pillow
